- Fixed Bug #14277: LinearFit incorrect when fitting data with error bars. - Added a new weighting method when fitting data: Direct Weighting.
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1040 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ qtiplot/src/analysis/fit_gsl.h
@@ -8,8 +8,8 @@ class Fit;
 	
 //! Structure for fitting data
 struct FitData {
-  size_t n;// number of points to be fitted (size of X, Y and sigma arrays)
-  size_t p;// number of fit parameters
+  int n;// number of points to be fitted (size of X, Y and sigma arrays)
+  int p;// number of fit parameters
   double * X;// the data to be fitted (abscissae) 
   double * Y; // the data to be fitted (ordinates)
   double * sigma; // the weighting data
